Day_06_Full_DDPM_Training_Loop/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
Pillow>=8.3.0
tqdm>=4.62.0
pyyaml>=6.0
wandb>=0.13.0
lpips>=0.1.4
scikit-image>=0.19.0
opencv-python>=4.5.0
tensorboard>=2.8.0
pytest>=7.0.0
jupyter>=1.0.0
ipykernel>=6.0.0